OT111-103 - Moved toEntity & refreshData methods to OrganizationRequest.
